Generate features in an json on postinstall
Close #18
